Distcount.Counters.TimeBucketAggregator (distcount v0.1.0)

Time bucket aggregator for counters.

The main idea of this module is to aggregate counter logs by time slot, where the time slot is the offloading interval (interval for offloading the counter logs to the database to sync state).

Since the counters are kept in memory (via an ETS table), the idea is to offload to the database only the counters that have been updated during the last offloading interval. Otherwise, we will end up having a full copy of the counters in memory and maybe offloading to the database some counters that haven't been even updated. This will be very inefficient and will affect the offloading performance significantly as well as it will increase the memory consumption very much (depending on the number of counters).

Usage

You can provide your own configuration via config file like so:

config :distcount, Distcount.Counters.TimeBucketAggregator,
  offload_interval: 5000

Check the available options in the start_link/1 function.

Telemetry

This module exposes following Telemetry events:

[:counters, :aggregator, :start] - Dispatched when the process starts.

Measurement: %{system_time: integer}

Metadata: %{name: atom}

[:counters, :aggregator, :stop] - Dispatched when the process stops.

Measurement: %{system_time: integer}

Metadata: %{name: atom, reason: term}

[:counters, :aggregator, :incr, :start] - Dispatched before the incr/3 function is executed.

Measurement: %{system_time: integer}

Metadata: %{counter: binary, amount: integer}

[:counters, :aggregator, :incr, :stop] - Dispatched after the incr/3 function is executed.

Measurement: %{duration: integer}

Metadata: %{counter: binary, amount: integer}

[:counters, :aggregator, :incr, :exception] - This event should be invoked when an error or exception occurs while executing the incr/3 function.

Measurement: %{duration: integer}

Metadata:

%{
  counter: binary,
  amount: integer,
  kind: :error | :exit | :throw,
  reason: term,
  stacktrace: term
}

[:counters, :aggregator, :offload, :start] - Dispatched before the offloading process is executed.

Measurement: %{system_time: integer}

Metadata: %{state: state}

[:counters, :aggregator, :offload, :stop] - Dispatched after the offloading process is executed.

Measurement: %{duration: integer}

Metadata: %{state: state}

[:counters, :aggregator, :offload, :exception] - This event should be invoked when an error or exception occurs while executing the offloading.

Measurement: %{duration: integer}

Metadata:

%{
  state: state,
  kind: :error | :exit | :throw,
  reason: term,
  stacktrace: term
}

[:counters, :aggregator, :offloaded_logs] - Dispatched after inserting the counter logs (offloading process).

Measurement: %{system_time: integer}

Metadata: %{inserted: integer, state: state}

Types

state()

Specs

state() :: %Distcount.Counters.TimeBucketAggregator{
  name: atom() | nil,
  offload_interval: non_neg_integer(),
  offload_timer_ref: :timer.tref() | nil,
  pid: pid() | nil,
  start_time: non_neg_integer() | nil,
  tid: :ets.tid() | nil
}

Type for internal state

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

incr(name_or_pid, counter, amount)

Specs

incr(
  name_or_pid :: atom() | pid(),
  counter :: binary(),
  amount :: pos_integer()
) :: pos_integer()

Increments/decrements the counter by the given amount.

Example

Distcount.Counters.TimeBucketAggregator.incr(:test, "counter", 1)

start_link(opts \\ [])

Specs

start_link(opts :: keyword()) :: GenServer.on_start()

Starts a new server for the time-bucket defined by the given options opts.

Options

:name - An atom defining the name of the server. Defaults __MODULE__.

:offload_interval - Offloading interval in milliseconds (Defaults to 10_000 - 10 sec).

Example

Distcount.Counters.TimeBucketAggregator.start_link(name: :test)